Fix deadband range for vertical overlap case
Previously the deadband range for transmission was too narrow
for the case where Q=0 on the downstream demand curve occurs
at a vertical portion of the curve (that is, where there is a
vertical step exactly at Q=0).  This case doesn't occur in the
test simulations and may not ever occur in practice.