Extend net demand curves upward when aggregating
Previously the sum of two curves was undefined for prices
above the first curve to end.  E.g., if curve 1 ended at $50
and curve 2 continued up to $80, the sum of the curves was
undefined after $50.  Now assume that the shorter curve extends
up as needed at the last value of q_min.  That is, assume
the curve is vertical above its last value of q_min on either
the supply or demand side.  Reduces the number of occasions
when no equilibrium exists.