Support logging fake demands; make getBid private.
Now writes forged demands to the net demand file just after
the true versions.

Switch getBid back to private, as it was originally, and create
two new methods, getBidMin() and getBidMax() to return the end
points of the step. Needed to keep Bidstep fully internal.
Also, eliminate obsolete p field from Bidstep and add logging
of the number of demands aggregated when debugging is on.